Refactor code to work with NextJS transpile packages
