+++ apps/app/.env.example
@@ -1,8 +1,5 @@
 # Stack Auth keys
 # Get these variables by creating a project on https://app.stack-auth.com.
-NEXT_PUBLIC_STACK_PROJECT_ID=your-next_public_stack_project_id
-NEXT_PUBLIC_STACK_PUBLISHABLE_CLIENT_KEY=your-next_public_stack_publishable_client_key
-STACK_SECRET_SERVER_KEY=your-stack_secret_server_key
+NEXT_PUBLIC_STACK_PROJECT_ID=your-stack_project_id
+NEXT_PUBLIC_STACK_PUBLISHABLE_CLIENT_KEY=your-stack_publishable_client_key
 
-NEXT_PUBLIC_POSTHOG_KEY=your-next_public_posthog_key
-NEXT_PUBLIC_POSTHOG_HOST=your-next_public_posthog_host
