+++ apps/app/src/app/(auth)/settings/page.tsx
@@ -1,11 +1,17 @@
-import { UserSettings } from "./UserSettings";
-import { RequireAuth, StackProvider, stackApp } from "@/components";
+import { AccountSettings, StackProvider } from "@stackframe/react";
+import { stackApp } from "~/lib/auth";
+import { Suspense } from "react";
+import { Loading } from "@hebo/ui/components/base/Loading";
 
 export default function Settings() {
     return (
         <StackProvider app={stackApp}>
+
             {/* <RequireAuth /> */}
-            <UserSettings />
+            <Suspense fallback={<Loading size="md" variant="primary" />}>
+                <AccountSettings fullPage={false} />
+            </Suspense>
+
         </StackProvider>
     );
 }
