--- apps/app/src/components/CodeSnippetWrapper.tsx
@@ -1,11 +1,0 @@
-"use client";
-
-import { CodeSnippet } from "@hebo/ui";
-
-interface CodeSnippetWrapperProps {
-  c: string;
-}
-
-export function CodeSnippetWrapper({ c }: CodeSnippetWrapperProps) {
-  return <CodeSnippet c={c} />;
-}
