--- apps/app/src/components/index.ts
@@ -1,6 +1,0 @@
-export { AuthUserButton } from "./AuthUserButton";
-export { Logo } from "./Logo";
-export { UserDisplay } from "./UserDisplay";
-export { RequireAuth } from "./functional/RequireAuth";
-export { stackApp, StackProvider } from "./functional/Stack";
-export { CodeSnippetWrapper } from "./CodeSnippetWrapper";
