+++ apps/app/src/components/ui/UserName.tsx
@@ -1,0 +1,13 @@
+"use client";
+
+//import { useStackApp } from "@stackframe/react";
+
+export function UserName() {
+  //const app = useStackApp();
+  //const user = app?.useUser();
+  
+  return (
+    //<span>{user?.displayName || 'User'}</span>
+    <span>User</span>
+  );
+}
