+++ apps/app/tailwind.config.ts
@@ -1,12 +1,12 @@
 import type { Config } from "tailwindcss";
-import { tailwindPreset } from "@hebo/ui";
+import preset from "./tailwind.preset";
 
 const config: Config = {
   content: [
     "./src/**/*.{js,jsx,ts,tsx}",
     "../../packages/ui/src/**/*.{js,jsx,ts,tsx}",
   ],
-  presets: [tailwindPreset],
+  presets: [preset],
 };
 
 export default config;
