+++ apps/app/tsconfig.json
@@ -3,14 +3,15 @@
   "compilerOptions": {
     "baseUrl": ".",
     "paths": {
-      "@/*": ["./src/*"]
+      "~/*": [ "./src/*" ]
     },
     "plugins": [
       {
         "name": "next"
       }
     ],
-    "strictNullChecks": true
+    "strictNullChecks": true,
+    "jsx": "preserve"
   },
   "include": [
     "next-env.d.ts",
@@ -19,5 +20,7 @@
     ".next/types/**/*.ts",
     "../../packages/*/src/**/*"
   ],
-  "exclude": ["node_modules"]
-} 
+  "exclude": [
+    "node_modules"
+  ]
+}
