+++ packages/ui/package.json
@@ -2,29 +2,20 @@
   "name": "@hebo/ui",
   "version": "0.0.1",
   "private": true,
-  "sideEffects": false,
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "files": [
-    "dist/**"
-  ],
+  "type": "module",
+  "sideEffects": true,
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
   "scripts": {
-    "build": "tsup src/index.ts --format cjs,esm --dts",
-    "dev": "tsup src/index.ts --format cjs,esm --dts --watch",
-    "lint": "eslint . --max-warnings 0",
-    "clean": "rm -rf .turbo node_modules dist"
+    "lint": "eslint . --max-warnings 0"
   },
   "peerDependencies": {
     "react": "*",
-    "react-dom": "*",
-    "tailwindcss": "*"
+    "react-dom": "*"
   },
   "devDependencies": {
     "@types/react": "*",
-    "@types/react-dom": "*",
-    "tailwindcss": "*",
-    "tailwindcss-animate": "*",
-    "tsup": "*"
+    "@types/react-dom": "*"
   },
   "dependencies": {
     "@radix-ui/react-dialog": "^1.1.14",
@@ -32,21 +23,8 @@
     "class-variance-authority": "*",
     "clsx": "*",
     "lucide-react": "*",
-    "next-themes": "^0.4.6",
     "sonner": "^2.0.5",
     "tailwind-merge": "*",
-    "tw-animate-css": "*",
     "vaul": "^1.1.2"
-  },
-  "exports": {
-    ".": "./src/index.ts",
-    "./globals.css": "./src/globals.css",
-    "./tailwind.preset": "./tailwind.preset.ts",
-    "./shadcn/button": "./src/shadcn/button.tsx",
-    "./shadcn/dialog": "./src/shadcn/dialog.tsx",
-    "./shadcn/drawer": "./src/shadcn/drawer.tsx",
-    "./shadcn/sonner": "./src/shadcn/sonner.tsx",
-    "./components/": "./src/components/",
-    "./lib/utils": "./src/lib/utils.ts"
   }
 }
