--- packages/ui/postcss.config.mjs
@@ -1,8 +1,0 @@
-const config = {
-  plugins: {
-    tailwindcss: {},
-    autoprefixer: {},
-  },
-};
-
-export default config;
