--- packages/ui/src/components/index.ts
@@ -1,4 +1,0 @@
-export { CodeSnippet } from "./CodeSnippet";
-export { ActionButton } from "./ActionButton";
-export { Loading } from "./dumb/loading";
-export { Sonner } from "./dumb/sonner";
