--- packages/ui/src/index.ts
@@ -1,3 +1,0 @@
-export * from "./lib/utils";
-export * from "./components";
-export { default as tailwindPreset } from "../tailwind.preset";
