--- packages/ui/tsup.config.ts
@@ -1,15 +1,0 @@
-import { defineConfig } from "tsup";
-
-export default defineConfig({
-  entry: ["src/index.ts"],
-  format: ["cjs", "esm"],
-  dts: true,
-  splitting: false,
-  sourcemap: true,
-  clean: true,
-  treeshake: true,
-  external: ["react", "react-dom"],
-  esbuildOptions(options) {
-    options.keepNames = true;
-  },
-});
