+++ tsconfig.base.json
@@ -13,15 +13,11 @@
     "moduleResolution": "bundler",
     "resolveJsonModule": true,
     "isolatedModules": true,
-    "jsx": "preserve",
+    "jsx": "react-jsx",
     "incremental": false,
     "baseUrl": ".",
     "paths": {
-      "@/*": ["*"],
-      "@hebo/*": ["packages/*"],
-      "@hebo/ui": ["packages/ui/"],
-      "@hebo/db": ["packages/db/"],
-      "@hebo/db/*": ["packages/db/*/"]
+      "@hebo/ui/*": ["packages/ui/src/*"]
     }
   },
   "exclude": ["node_modules"]
